refactor: migrate nodeBase mixin to composable
